+++ app-backend/api/src/main/scala/project/ProjectLayerAnnotationRoutes.scala
@@ -5,10 +5,12 @@ import com.rasterfoundry.datamodel._
 import com.rasterfoundry.datamodel.GeoJsonCodec._
 import com.rasterfoundry.database._
 import com.rasterfoundry.akkautil._
+import com.rasterfoundry.common.AWSBatch
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server._
 import cats.effect._
+import cats.implicits._
 import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 import doobie.Transactor
 import doobie.implicits._
@@ -22,7 +24,8 @@ trait ProjectLayerAnnotationRoutes
     with CommonHandlers
     with PaginationDirectives
     with ProjectAuthorizationDirectives
-    with QueryParametersCommon {
+    with QueryParametersCommon
+    with AWSBatch {
 
   implicit val xa: Transactor[IO]
 
@@ -380,5 +383,178 @@ trait ProjectLayerAnnotationRoutes
         }
       }
     }
+
+  def listGeojsonUploads(projectId: UUID,
+                         layerId: UUID,
+                         annotationGroupId: UUID): Route =
+    authenticate { user =>
+      withPagination { page =>
+        authorizeAuthResultAsync {
+          ProjectDao
+            .authorized(
+              user,
+              ObjectType.Project,
+              projectId,
+              ActionType.Annotate
+            )
+            .transact(xa)
+            .unsafeToFuture
+        } {
+          complete {
+            GeojsonUploadDao
+              .listUploadsForAnnotationGroup(projectId,
+                                             layerId,
+                                             annotationGroupId,
+                                             page)
+              .transact(xa)
+              .unsafeToFuture
+          }
+        }
+      }
+    }
+
+  def createGeojsonUpload(
+      projectId: UUID,
+      layerId: UUID,
+      annotationGroupId: UUID
+  ): Route =
+    authenticate { user =>
+      authorizeAuthResultAsync {
+        ProjectDao
+          .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
+          .transact(xa)
+          .unsafeToFuture
+      } {
+        entity(as[GeojsonUpload.Create]) { newUpload =>
+          logger.debug(
+            s"newUpload: ${newUpload.uploadType}, ${newUpload.files}, ${newUpload.fileType}"
+          )
+          val uploadToInsert =
+            (newUpload.uploadType, newUpload.files.nonEmpty, newUpload.fileType) match {
+              case (UploadType.S3, true, FileType.GeoJson) => {
+                newUpload
+              }
+              case (_, false, _) => {
+                throw new IllegalStateException(
+                  "Uploads must specify file URIs"
+                )
+              }
+              case (_, _, fileType) if fileType.equals(FileType.GeoJson) => {
+                throw new IllegalStateException(
+                  "Unsupported file type. Please use GeoJson"
+                )
+              }
+              case _ => {
+                throw new IllegalStateException(
+                  "Non-S3 uploads currently unsupported"
+                )
+              }
+            }
+
+          onSuccess({
+            val uploadIO = AnnotationGroupDao
+              .getAnnotationGroup(projectId, annotationGroupId)
+              .flatMap {
+                _ traverse { _ =>
+                  GeojsonUploadDao
+                    .insert(
+                      uploadToInsert,
+                      projectId,
+                      layerId,
+                      annotationGroupId,
+                      user
+                    )
+                }
+              }
+            uploadIO.transact(xa).unsafeToFuture
+          }) { uploadO =>
+            uploadO match {
+              case Some(upload) =>
+                if (upload.uploadStatus.equals(UploadStatus.Uploaded)) {
+                  kickoffGeojsonImport(upload.id)
+                }
+                complete((StatusCodes.Created, upload))
+              case _ =>
+                complete(StatusCodes.NotFound)
+            }
+          }
+        }
+      }
+    }
+
+  def getGeojsonUpload(
+      projectId: UUID,
+      layerId: UUID,
+      annotationGroupId: UUID,
+      uploadId: UUID
+  ): Route = authenticate { user =>
+    authorizeAuthResultAsync {
+      ProjectDao
+        .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
+        .transact(xa)
+        .unsafeToFuture
+    } {
+      complete {
+        GeojsonUploadDao
+          .getLayerUpload(projectId, layerId, annotationGroupId, uploadId)
+          .transact(xa)
+          .unsafeToFuture
+      }
+    }
+  }
+
+  def updateGeojsonUpload(
+      projectId: UUID,
+      projectLayerId: UUID,
+      annotationGroupId: UUID,
+      uploadId: UUID
+  ): Route = authenticate { user =>
+    authorizeAuthResultAsync {
+      ProjectDao
+        .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
+        .transact(xa)
+        .unsafeToFuture
+    } {
+      entity(as[GeojsonUpload]) { upload =>
+        complete {
+          GeojsonUploadDao
+            .update(upload,
+                    projectId,
+                    projectLayerId,
+                    annotationGroupId,
+                    uploadId,
+                    user)
+            .transact(xa)
+            .unsafeToFuture
+        }
+      }
+    }
+  }
+
+  def deleteGeojsonUpload(
+      projectId: UUID,
+      layerId: UUID,
+      annotationGroupId: UUID,
+      uploadId: UUID
+  ): Route = authenticate { user =>
+    authorizeAuthResultAsync {
+      ProjectDao
+        .authorized(user, ObjectType.Project, projectId, ActionType.Annotate)
+        .transact(xa)
+        .unsafeToFuture
+    } {
+      onSuccess(
+        GeojsonUploadDao
+          .deleteProjectLayerUpload(projectId,
+                                    layerId,
+                                    annotationGroupId,
+                                    uploadId)
+          .transact(xa)
+          .unsafeToFuture
+      ) {
+        completeSingleOrNotFound
+      }
+    }
+  }
 
 }