+++ app-backend/common/src/main/scala/AWSBatch.scala
@@ -75,6 +75,17 @@ trait AWSBatch extends RollbarNotifier with LazyLogging {
                      jobName)
   }
 
+  def kickoffGeojsonImport(uploadId: UUID): Unit = {
+    val jobDefinition = awsbatchConfig.geojsonImportJobName
+    val jobName = s"$jobDefinition-$uploadId"
+    submitJobRequest(
+      jobDefinition,
+      awsbatchConfig.jobQueue,
+      Map("uploadId" -> s"$uploadId"),
+      jobName
+    )
+  }
+
   def kickoffProjectExport(exportId: UUID): Unit = {
     val jobDefinition = awsbatchConfig.exportJobName
     val jobName = s"$jobDefinition-$exportId"