+++ app-backend/datamodel/src/main/scala/FileType.scala
@@ -11,11 +11,13 @@ object FileType {
   case object Geotiff extends FileType("GEOTIFF")
   case object GeotiffWithMetadata extends FileType("GEOTIFF_WITH_METADATA")
   case object NonSpatial extends FileType("NON_SPATIAL")
+  case object GeoJson extends FileType("GEOJSON")
 
   def fromString(s: String): FileType = s.toUpperCase match {
     case "GEOTIFF"               => Geotiff
     case "GEOTIFF_WITH_METADATA" => GeotiffWithMetadata
     case "NON_SPATIAL"           => NonSpatial
+    case "GEOJSON"               => GeoJson
   }
 
   implicit val fileTypeEncoder: Encoder[FileType] =